python-rag-service/requirements.txt
# Core
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PDF Processing
pymupdf==1.23.21  # (fitz) - Best PDF text extraction
pdfplumber==0.10.3  # Fallback PDF processor
pypdf2==3.0.1  # Alternative PDF library

# AI/Vector/LLM
openai>=1.25.0
langchain-openai>=0.1.6
langchain==0.1.11
faiss-cpu==1.7.4  # Use faiss-gpu if you have CUDA
sentence-transformers==2.2.2
onnxruntime==1.16.3  # For your reranking

# Text Processing
nltk==3.8.1
numpy==1.26.3
scikit-learn==1.3.2  # For BM25 implementation

# Utilities
tqdm==4.66.1  # Progress bars
python-dateutil==2.8.2
pydantic==2.5.3
pydantic-settings==2.1.0
pytesseract==0.3.10


# Optional (for evaluation)
scipy==1.11.4  # For advanced metrics